+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## [Unreleased]
 
-### Added
 - feat: Add `phoneFormatProperties` field to Phone fields returning detailed format information (label, mask, regex, instruction, type). Resolves #441.
+- fix: Correctly resolve `AddressField.defaultState`, `AddressField.defaultProvince` based on the `addressType`. H/t @byanko-bot
+- fix: Set the default `AddressField.inputs.label` for the State/Province and Zip/Postal inputs based on the `addressType`. H/t @byanko-bot
+- chore: Update Composer and NPM deps.
 
 ## [v0.13.3]